리뷰 등록, 조회

+++ src/main/kotlin/com/brandjunhoe/productservice/review/application/dto/ReviewTotalDTO.kt
@@ -5,7 +5,7 @@ package com.brandjunhoe.productservice.review.application.dto
  */
 class ReviewTotalDTO(
 
-    val scoreAvg: Double,
+    val scoreAvg: Float,
 
     val count: Int
